Split the sequencer file into a file per class
